move getCollisionType to snakeHelpers
